requirements.txt
Django==5.0.3
djangorestframework==3.15.1
requests==2.31.0
pkce==1.0.3
Pillow==10.3.0
django-cors-headers==4.3.1
distutils-pytest==0.2.1
sphinx==7.3.7
mysqlclient